refactor: optimize code structure
